fix creators re-ordering bug
